Compilation fix for periodPWM variable

+++ hardware/lm4f/cores/lm4f/wiring_analog.c
@@ -105,7 +105,7 @@ void PWMWrite(uint8_t pin, uint32_t analog_res, uint32_t duty, unsigned int freq
 
         if (port == NOT_A_PORT) return; 	// pin on timer?
 
-        uint64_t periodPWM = ROM_SysCtlClockGet()/freq;
+        uint32_t periodPWM = ROM_SysCtlClockGet()/freq;
 
         enableTimerPeriph(offset);
         ROM_GPIOPinConfigure(timerToPinConfig(timer));
